Fix NSUInteger and int convert issue in ZXCVBN algorithm

+++ PassPalette/CPRectLayout.m
@@ -25,7 +25,7 @@
 - (void)prepareLayout {
     NSMutableArray *mutableLayoutAttribute = [NSMutableArray array];
     
-    int count = [self.collectionView numberOfItemsInSection:0];
+    NSUInteger count = [self.collectionView numberOfItemsInSection:0];
     static const CGFloat topSeperator = 20.0;
     static const CGFloat seperator = 4.0;
     CGFloat width = (self.collectionViewContentSize.width - seperator) / 2;
